requirements_part1.txt
sklearn
pandas
numpy
rdkit-pypi
torch